resturcture and componentize code
